requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==2.4.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudflare==3.1.0
colorama==0.4.6
cron-descriptor==1.4.5
decorator==5.1.1
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.3
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-tenants==3.7.0
django-tenants-celery-beat==0.2.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
inject==5.2.1
injector==0.22.0
ipdb==0.13.13
ipython==8.29.0
jedi==0.19.2
kombu==5.4.2
matplotlib-inline==0.1.7
parso==0.8.4
pillow==11.1.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
ruff==0.8.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stack-data==0.6.3
tenant-schemas-celery==3.0.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
